Flow for add car

+++ commons/src/main/java/com/fortech/Car.java
@@ -3,15 +3,15 @@ package com.fortech;
 import java.io.Serializable;
 import java.util.Date;
 
+
 public class Car implements Serializable {
 
     private String name;
     private String mark;
     private String color;
-    private float price;
+    private double price;
     private String condition;
     private Date registrationDate;
-   //private Tim
 
     public String getName() {
         return name;
@@ -37,11 +37,11 @@ public class Car implements Serializable {
         this.color = color;
     }
 
-    public float getPrice() {
+    public double getPrice() {
         return price;
     }
 
-    public void setPrice(float price) {this.price = price; }
+    public void setPrice(double price) {this.price = price; }
 
     public String getCondition() {
         return condition;
